+++ src/run.rs
@@ -589,6 +589,12 @@ code:
     pub fn span(&self) -> Span {
         self.spans.lock()[self.span_index()].clone()
     }
+    pub(crate) fn merge_spans(&self, a: usize, b: usize) -> Span {
+        let spans = self.spans.lock();
+        let a = spans[a].clone();
+        let b = spans[b].clone();
+        a.merge(b)
+    }
     /// Construct an error with the current span
     pub fn error(&self, message: impl ToString) -> UiuaError {
         UiuaError::Run(self.span().clone().sp(message.to_string()))
@@ -637,6 +643,9 @@ code:
     }
     pub fn diagnostics(&self) -> &BTreeSet<Diagnostic> {
         &self.diagnostics
+    }
+    pub fn diagnostics_mut(&mut self) -> &mut BTreeSet<Diagnostic> {
+        &mut self.diagnostics
     }
     pub fn take_diagnostics(&mut self) -> BTreeSet<Diagnostic> {
         take(&mut self.diagnostics)