+++ changelog.md
@@ -2,6 +2,14 @@
 
 Uiua is not yet stable.
 
+## 0.0.18 - 2023-10-??
+### Language
+- [`if` `?`](https://uiua.org/docs/if)'s branches can now have a different number of arguments (but not outputs)
+- [`under` `⍜`](https://uiua.org/docs/under) [`&n`](https://uiua.org/docs/&n) can be used to time things
+### Interpreter
+- The formatter now aligns consecutive end-of-line comments
+- Fix a bug that prevent [`under` `⍜`](https://uiua.org/docs/under) multidimensional [`take` `↙`](https://uiua.org/docs/take) and [`drop` `↘`](https://uiua.org/docs/drop) from working
+
 ## 0.0.17 - 2023-10-07
 ### Language
 - Add GIF encoding with [`&gife`](https://uiua.org/docs/&gife)